// src/app/swagger-models/model/customer.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */import { Address } from './address';
import { CreditCard } from './creditCard';


export interface Customer { 
    address?: Address;
    creditCard?: CreditCard;
    email?: string;
    firstName?: string;
    id?: number;
    lastName?: string;
}