// src/app/swagger-models/model/salesOrder.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */import { Address } from './address';
import { CreditCard } from './creditCard';
import { CustomerInfo } from './customerInfo';
import { SalesOrderItem } from './salesOrderItem';


export interface SalesOrder { 
    address?: Address;
    amount?: number;
    creditCard?: CreditCard;
    customer?: CustomerInfo;
    date?: Date;
    id?: number;
    items?: Array<SalesOrderItem>;
    status?: SalesOrder.StatusEnum;
}
export namespace SalesOrder {
    export type StatusEnum = 'ACCEPTED' | 'PROCESSING' | 'SHIPPED' | 'CANCELED';
    export const StatusEnum = {
        ACCEPTED: 'ACCEPTED' as StatusEnum,
        PROCESSING: 'PROCESSING' as StatusEnum,
        SHIPPED: 'SHIPPED' as StatusEnum,
        CANCELED: 'CANCELED' as StatusEnum
    };
}