// src/app/swagger-models/model/salesOrderItem.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BookInfo } from './bookInfo';


export interface SalesOrderItem { 
    book?: BookInfo;
    price?: number;
    quantity?: number;
}